Updated at 5/13/25 16:37 经理API基本做完

+++ src/main/java/com/winniethepooh/hotelsystembackend/vo/DishTop10VO.java
@@ -2,10 +2,8 @@ package com.winniethepooh.hotelsystembackend.vo;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class DishTop10VO {
     private String name;
-    private BigDecimal value;
+    private Integer value;
 }
